refactor(scenario): split load-time tree-hydration repair into ParsekScenario.HydrationRepair.cs
